Update README.md
In this example, the conversation rates from USD to EUR and GBP are hardcoded asIn this example, the conversion rates from USD to EUR and GBP are hardcoded as usd_to_eur and usd_to_gbp, respectively. You can modify these values to use the most up-to-date conversion rates.

The program prompts the user to enter an amount in US dollars, then calculates and displays the equivalent amounts in euros and British pounds.

Please note that currency exchange rates change frequently, so make sure to update the conversion rates accordingly to get accurate results.